move static vector types to smatrix
