+++ tests/matrix_ops.rs
@@ -584,9 +584,9 @@ mod dmatrix_tests
     use approx::assert_relative_eq;
     use topohedral_linalg::dmatrix::*;
     use topohedral_linalg::dvector::{DVector, VecType};
-    use topohedral_linalg::scvector::SCVector;
+    use topohedral_linalg::smatrix::SCVector;
     use topohedral_linalg::smatrix::SMatrix;
-    use topohedral_linalg::srvector::SRVector;
+    use topohedral_linalg::smatrix::SRVector;
     use topohedral_linalg::{Complex, MatMul, MatrixOps};
 
     //{{{ collection: eig tests
